theme: make UI2 the default theme (#103481)
Fix DE-370
