+++ static/app/views/nav/index.spec.tsx
@@ -514,181 +514,4 @@ describe('Nav', () => {
       expect(screen.queryByRole('dialog')).not.toBeInTheDocument();
     });
   });
-
-  describe('chonk-ui', () => {
-    describe('switching themes', () => {
-      beforeEach(() => {
-        MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/prompts-activity/',
-          body: {data: {dismissed_ts: null}},
-        });
-
-        ConfigStore.set('user', {
-          ...ConfigStore.get('user'),
-          options: {
-            ...ConfigStore.get('user').options,
-            prefersChonkUI: false,
-          },
-        });
-      });
-
-      describe('when feature flag is enabled', () => {
-        it('shows the chonk-ui toggle in the help menu', async () => {
-          const dismissRequest = MockApiClient.addMockResponse({
-            url: '/organizations/org-slug/prompts-activity/',
-            method: 'PUT',
-          });
-
-          renderNav({features: ALL_AVAILABLE_FEATURES.concat('chonk-ui')});
-          const helpMenu = screen.getByRole('button', {name: 'Help'});
-          await userEvent.click(helpMenu);
-
-          expect(screen.getByText('Try our new look')).toBeInTheDocument();
-
-          // Once for banner, once for dot indicator
-          expect(dismissRequest).toHaveBeenCalledTimes(2);
-        });
-
-        it('shows the chonk-ui toggle to old theme', async () => {
-          ConfigStore.set('user', {
-            ...ConfigStore.get('user'),
-            options: {
-              ...ConfigStore.get('user').options,
-              prefersChonkUI: true,
-            },
-          });
-
-          renderNav({features: ALL_AVAILABLE_FEATURES.concat('chonk-ui')});
-          const helpMenu = screen.getByRole('button', {name: 'Help'});
-          await userEvent.click(helpMenu);
-
-          expect(screen.getByText('Switch back to our old look')).toBeInTheDocument();
-        });
-      });
-
-      describe('when feature flag is disabled', () => {
-        it('does not show the chonk-ui toggle in the help menu', async () => {
-          MockApiClient.addMockResponse({
-            url: '/organizations/org-slug/prompts-activity/',
-            body: {data: {dismissed_ts: null}},
-          });
-
-          renderNav({features: ALL_AVAILABLE_FEATURES});
-          const helpMenu = screen.getByRole('button', {name: 'Help'});
-          await userEvent.click(helpMenu);
-
-          expect(screen.queryByText('Try our new look')).not.toBeInTheDocument();
-        });
-      });
-    });
-
-    describe('opt-in banner', () => {
-      it('shows the opt-in banner if user has feature and has not opted in yet', async () => {
-        MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/prompts-activity/',
-          body: {data: {dismissed_ts: null}},
-        });
-
-        renderNav({features: ALL_AVAILABLE_FEATURES.concat('chonk-ui')});
-        expect(await screen.findByText(/Sentry has a new look/)).toBeInTheDocument();
-      });
-
-      it('dismissing the banner', async () => {
-        MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/prompts-activity/',
-          body: {data: {dismissed_ts: null}},
-        });
-
-        const dismissRequest = MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/prompts-activity/',
-          method: 'PUT',
-          status: 200,
-        });
-
-        renderNav({features: ALL_AVAILABLE_FEATURES.concat('chonk-ui')});
-        expect(await screen.findByText(/Sentry has a new look/)).toBeInTheDocument();
-
-        await userEvent.click(screen.getByRole('button', {name: 'Dismiss'}));
-
-        expect(dismissRequest).toHaveBeenCalled();
-
-        await waitFor(() => {
-          expect(dismissRequest).toHaveBeenCalledWith(
-            '/organizations/org-slug/prompts-activity/',
-            expect.objectContaining({
-              method: 'PUT',
-              data: expect.objectContaining({
-                feature: 'chonk_ui_banner',
-                status: 'snoozed',
-              }),
-            })
-          );
-        });
-
-        expect(screen.queryByText(/Sentry has a new look/)).not.toBeInTheDocument();
-      });
-
-      it('enabling new theme dismisses banner and dot indicator', async () => {
-        MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/prompts-activity/',
-          body: {data: {dismissed_ts: null}},
-        });
-
-        const optInRequest = MockApiClient.addMockResponse({
-          url: '/users/me/',
-          method: 'PUT',
-        });
-
-        const dismissRequest = MockApiClient.addMockResponse({
-          url: '/organizations/org-slug/prompts-activity/',
-          method: 'PUT',
-        });
-
-        renderNav({features: ALL_AVAILABLE_FEATURES.concat('chonk-ui')});
-        expect(await screen.findByText(/Sentry has a new look/)).toBeInTheDocument();
-
-        // Enables user option and disables all prompts
-        await userEvent.click(screen.getByText('Try It Out'));
-
-        expect(optInRequest).toHaveBeenCalledWith(
-          '/users/me/',
-          expect.objectContaining({
-            method: 'PUT',
-            data: expect.objectContaining({
-              options: expect.objectContaining({
-                prefersChonkUI: true,
-              }),
-            }),
-          })
-        );
-
-        expect(dismissRequest).toHaveBeenNthCalledWith(
-          1,
-          '/organizations/org-slug/prompts-activity/',
-          expect.objectContaining({
-            method: 'PUT',
-            data: expect.objectContaining({
-              feature: 'chonk_ui_banner',
-              status: 'snoozed',
-            }),
-          })
-        );
-
-        expect(dismissRequest).toHaveBeenNthCalledWith(
-          2,
-          '/organizations/org-slug/prompts-activity/',
-          expect.objectContaining({
-            method: 'PUT',
-            data: expect.objectContaining({
-              feature: 'chonk_ui_dot_indicator',
-              status: 'snoozed',
-            }),
-          })
-        );
-
-        // The banner is no longer visible
-        expect(screen.queryByText(/Sentry has a new look/)).not.toBeInTheDocument();
-      });
-    });
-  });
 });
